+++ CHANGELOG.md
@@ -8,9 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.8.0] -
 
 * Created `MultiSelectDelegate` and `MultiSelectCollectionViewController` based on [objc.io's Generic Table View Controllers](https://talk.objc.io/episodes/S01E26-generic-table-view-controllers-part-2)
+* Added `ItemsViewController` based on the same source as above
 * Made selectDelegate an optional in `FlaneurFormSelectElementCollectionViewCell`
 * Fixed cell selection management in FlaneurFormSelectElementCollectionViewCell #24
 * Changed design of form cells so that they don't use conditional/forced casts so much
+* Added code to create snapshots of a `MKMapView`
+* Added default values for MapKit types
+* Added code to render a grayscale version of a `UIImage`
 
 ## [0.7.0] - 2018-01-18